独行月球影评分析/Code/requirements.txt
tqdm==4.63.0
numpy==1.22.4
jieba==0.42.1
pandas==1.4.3
PySide6==6.3.1
requests==2.27.1
jupyterlab==3.4.5
matplotlib==3.5.2
wordcloud==1.8.2.2
paddlenlp==2.4.0
paddlepaddle==2.3.2
scikit-learn==0.23.2
beautifulsoup4==4.11.1
